backend/requirements.txt
requests==2.31.0
bs4==0.0.1
psycopg2-binary==2.9.9
httptools==0.6.1
python-dotenv==1.0.0
alembic==1.8.1
fastapi==0.115.12
pydantic==2.11.4
pytest==7.4.3
sqlalchemy==2.0.25
uvicorn==0.20.0
websockets==12.0
black==24.1.1
dependency-injector==4.41.0